Remove CreateCompletedCalculationActivity and related
